Changed Ad Classes to AdHandler Classes to prevent a naming conflict with Composable functions Changed Remember functions to load ads upon creation Moved all Ad Composable work into LaunchedEffect calls to ensure no perceived lag
