+++ basic-ads/src/androidMain/kotlin/app/lexilabs/basic/ads/InterstitialAdHandler.kt
@@ -1,46 +1,47 @@
 package app.lexilabs.basic.ads
 
 import android.app.Activity
+import androidx.annotation.MainThread
 import app.lexilabs.basic.logging.Log
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.LoadAdError
-import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAdLoadCallback
-import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAd as AndroidRewardedInterstitialAd
+import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback
+import com.google.android.gms.ads.interstitial.InterstitialAd as AndroidInterstitialAd
 
-public actual class RewardedInterstitialAd actual constructor(
+public actual class InterstitialAdHandler actual constructor(
     private val activity: Any?
 ) {
 
-    private val tag = "RewardedInterstitialAd"
-    private var rewardedInterstitialAd: AndroidRewardedInterstitialAd? = null
+    private val tag = "InterstitialAd"
+    private var interstitialAd: AndroidInterstitialAd? = null
 
     public actual fun load(
         adUnitId: String,
         onLoad: () -> Unit,
         onFailure: (Exception) -> Unit
-    ){
-        Log.d(tag, "loadRewardedAd: Loading")
+    ) {
+        Log.d(tag, "loadInterstitialAd: Loading")
         require(activity != null) {
             "Activity Context must be set to non-null value in Android"
         }
         require(activity is Activity) {
             "activity variable must be of the Android `Activity` type"
         }
-        AndroidRewardedInterstitialAd.load(
+        AndroidInterstitialAd.load(
             activity,
             adUnitId,
             AdRequest.Builder().build(),
-            object : RewardedInterstitialAdLoadCallback() {
+            object : InterstitialAdLoadCallback() {
                 override fun onAdFailedToLoad(adError: LoadAdError) {
                     super.onAdFailedToLoad(adError)
-                    Log.d(tag, "loadRewardedAd:failure:$adError")
+                    Log.d(tag, "loadInterstitialAd:failure:$adError")
                     onFailure(AdException(adError.message))
                 }
 
-                override fun onAdLoaded(ad: AndroidRewardedInterstitialAd) {
+                override fun onAdLoaded(ad: AndroidInterstitialAd) {
                     super.onAdLoaded(ad)
-                    Log.d(tag, "loadRewardedAd:success")
-                    rewardedInterstitialAd = ad
+                    Log.d(tag, "loadInterstitialAd:success")
+                    interstitialAd = ad
                     onLoad()
                 }
             }
@@ -55,21 +56,22 @@ public actual class RewardedInterstitialAd actual constructor(
         onClick: () -> Unit
     ) {
         Log.d(tag, "setListeners: Loading")
-        require(rewardedInterstitialAd != null) {
-            "RewardedAd not loaded yet. `RewardedAd.load()` must be called first"
+        require(interstitialAd != null) {
+            "InterstitialAd not loaded yet. `InterstitialAd.load()` must be called first"
         }
-        rewardedInterstitialAd?.let {
-            rewardedInterstitialAd?.fullScreenContentCallback = FullscreenContentDelegate(
+        interstitialAd?.let {
+            interstitialAd?.fullScreenContentCallback = FullscreenContentDelegate(
                 onClick = onClick,
                 onDismissed = onDismissed,
                 onFailure = onFailure,
                 onImpression = onImpression,
                 onShown = onShown
             )
-        } ?: Log.d(tag, "The rewarded interstitial ad wasn't ready yet.")
+        } ?: Log.d(tag, "The interstitial ad wasn't ready yet.")
     }
 
-    public actual fun show(onRewardEarned: () -> Unit) {
+    @MainThread
+    public actual fun show() {
         Log.d(tag, "show: Loading")
         require(activity != null) {
             "Activity Context must be set to non-null value in Android"
@@ -77,12 +79,10 @@ public actual class RewardedInterstitialAd actual constructor(
         require(activity is Activity) {
             "activity variable must be of the Android `Activity` type"
         }
-        require(rewardedInterstitialAd != null) {
-            "RewardedAd not loaded yet. `RewardedAd.load()` must be called first"
+        require(interstitialAd != null) {
+            "InterstitialAd not loaded yet. `InterstitialAd.load()` must be called first"
         }
-        rewardedInterstitialAd?.show(activity) {
-            Log.d(tag, "A reward was earned")
-            onRewardEarned()
-        }
+        interstitialAd?.show(activity)
     }
+
 }
