+++ basic-ads/src/androidMain/kotlin/app/lexilabs/basic/ads/RewardedAdHandler.kt
@@ -1,47 +1,44 @@
 package app.lexilabs.basic.ads
 
 import android.app.Activity
-import androidx.annotation.MainThread
 import app.lexilabs.basic.logging.Log
-import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.LoadAdError
-import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback
-import com.google.android.gms.ads.interstitial.InterstitialAd as AndroidInterstitialAd
+import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback
+import com.google.android.gms.ads.AdRequest as AndroidAdRequest
+import com.google.android.gms.ads.rewarded.RewardedAd as AndroidRewardedAd
 
-public actual class InterstitialAd actual constructor(
-    private val activity: Any?
-) {
+public actual class RewardedAdHandler actual constructor(private val activity: Any?) {
 
-    private val tag = "InterstitialAd"
-    private var interstitialAd: AndroidInterstitialAd? = null
+    private val tag = "RewardedAd"
+    private var rewardedAd: AndroidRewardedAd? = null
 
     public actual fun load(
         adUnitId: String,
         onLoad: () -> Unit,
         onFailure: (Exception) -> Unit
-    ) {
-        Log.d(tag, "loadInterstitialAd: Loading")
+    ){
+        Log.d(tag, "loadRewardedAd: Loading")
         require(activity != null) {
             "Activity Context must be set to non-null value in Android"
         }
         require(activity is Activity) {
             "activity variable must be of the Android `Activity` type"
         }
-        AndroidInterstitialAd.load(
+        AndroidRewardedAd.load(
             activity,
             adUnitId,
-            AdRequest.Builder().build(),
-            object : InterstitialAdLoadCallback() {
+            AndroidAdRequest.Builder().build(),
+            object : RewardedAdLoadCallback() {
                 override fun onAdFailedToLoad(adError: LoadAdError) {
                     super.onAdFailedToLoad(adError)
-                    Log.d(tag, "loadInterstitialAd:failure:$adError")
+                    Log.d(tag, "loadRewardedAd:failure:$adError")
                     onFailure(AdException(adError.message))
                 }
 
-                override fun onAdLoaded(ad: AndroidInterstitialAd) {
+                override fun onAdLoaded(ad: AndroidRewardedAd) {
                     super.onAdLoaded(ad)
-                    Log.d(tag, "loadInterstitialAd:success")
-                    interstitialAd = ad
+                    Log.d(tag, "loadRewardedAd:success")
+                    rewardedAd = ad
                     onLoad()
                 }
             }
@@ -54,24 +51,25 @@ public actual class InterstitialAd actual constructor(
         onShown: () -> Unit,
         onImpression: () -> Unit,
         onClick: () -> Unit
-    ) {
+    ){
         Log.d(tag, "setListeners: Loading")
-        require(interstitialAd != null) {
-            "InterstitialAd not loaded yet. `InterstitialAd.load()` must be called first"
+        require(rewardedAd != null) {
+            "RewardedAd not loaded yet. `RewardedAd.load()` must be called first"
         }
-        interstitialAd?.let {
-            interstitialAd?.fullScreenContentCallback = FullscreenContentDelegate(
+        rewardedAd?.let {
+            rewardedAd?.fullScreenContentCallback = FullscreenContentDelegate(
                 onClick = onClick,
                 onDismissed = onDismissed,
                 onFailure = onFailure,
                 onImpression = onImpression,
                 onShown = onShown
             )
-        } ?: Log.d(tag, "The interstitial ad wasn't ready yet.")
+        } ?: Log.d(tag, "The rewarded ad wasn't ready yet.")
     }
 
-    @MainThread
-    public actual fun show() {
+    public actual fun show(
+        onRewardEarned: () -> Unit
+    ){
         Log.d(tag, "show: Loading")
         require(activity != null) {
             "Activity Context must be set to non-null value in Android"
@@ -79,10 +77,12 @@ public actual class InterstitialAd actual constructor(
         require(activity is Activity) {
             "activity variable must be of the Android `Activity` type"
         }
-        require(interstitialAd != null) {
-            "InterstitialAd not loaded yet. `InterstitialAd.load()` must be called first"
+        require(rewardedAd != null) {
+            "RewardedAd not loaded yet. `RewardedAd.load()` must be called first"
         }
-        interstitialAd?.show(activity)
+        rewardedAd?.show(activity) {
+            Log.d(tag, "A reward was earned")
+            onRewardEarned()
+        }
     }
-
 }
