+++ basic-ads/src/commonMain/kotlin/app/lexilabs/basic/ads/composable/RememberInterstitialAd.kt
@@ -4,10 +4,23 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import app.lexilabs.basic.ads.AdUnitId
 import app.lexilabs.basic.ads.DependsOnGoogleMobileAds
-import app.lexilabs.basic.ads.InterstitialAd
+import app.lexilabs.basic.ads.InterstitialAdHandler
 
 @DependsOnGoogleMobileAds
 @Composable
-public fun rememberInterstitialAd(activity: Any?): MutableState<InterstitialAd> =
-    remember(activity) { mutableStateOf(InterstitialAd(activity)) }
+public fun rememberInterstitialAd(
+    activity: Any?,
+    adUnitId: String = AdUnitId.INTERSTITIAL_DEFAULT,
+    onLoad: () -> Unit = {},
+    onFailure: (Exception) -> Unit = {}
+): MutableState<InterstitialAdHandler> {
+    val ad = remember(activity) { mutableStateOf(InterstitialAdHandler(activity)) }
+    ad.value.load(
+        adUnitId = adUnitId,
+        onLoad = onLoad,
+        onFailure = onFailure
+    )
+    return ad
+}
