+++ basic-ads/src/commonMain/kotlin/app/lexilabs/basic/ads/composable/RememberRewardedAd.kt
@@ -4,10 +4,23 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import app.lexilabs.basic.ads.AdUnitId
 import app.lexilabs.basic.ads.DependsOnGoogleMobileAds
-import app.lexilabs.basic.ads.RewardedAd
+import app.lexilabs.basic.ads.RewardedAdHandler
 
 @DependsOnGoogleMobileAds
 @Composable
-public fun rememberRewardedAd(activity: Any?): MutableState<RewardedAd> =
-    remember(activity) { mutableStateOf(RewardedAd(activity)) }
+public fun rememberRewardedAd(
+    activity: Any?,
+    adUnitId: String = AdUnitId.REWARDED_DEFAULT,
+    onLoad: () -> Unit = {},
+    onFailure: (Exception) -> Unit = {}
+): MutableState<RewardedAdHandler> {
+    val ad = remember(activity) { mutableStateOf(RewardedAdHandler(activity)) }
+    ad.value.load(
+        adUnitId = adUnitId,
+        onLoad = onLoad,
+        onFailure = onFailure
+    )
+    return ad
+}
