+++ basic-ads/src/commonMain/kotlin/app/lexilabs/basic/ads/composable/RewardedAd.kt
@@ -1,11 +1,12 @@
 package app.lexilabs.basic.ads.composable
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import app.lexilabs.basic.ads.AdUnitId
 import app.lexilabs.basic.ads.DependsOnGoogleMobileAds
-import app.lexilabs.basic.ads.InterstitialAd
-import app.lexilabs.basic.ads.RewardedAd
+import app.lexilabs.basic.ads.InterstitialAdHandler
+import app.lexilabs.basic.ads.RewardedAdHandler
 
 /**
  * Loads and displays a Rewarded Ad using a [Composable].
@@ -17,7 +18,7 @@ import app.lexilabs.basic.ads.RewardedAd
  * @param onImpression Lambda expression that executes after the user has seen the ad
  * @param onClick Lambda expression that executes after the user clicks the ad
  * @param onFailure Lambda expression that executes after the ad fails to load or redirect
- * @param onLoad Lambda expression that executes after the [InterstitialAd] has fully loaded
+ * @param onLoad Lambda expression that executes after the [InterstitialAdHandler] has fully loaded
  * @see AdUnitId.autoSelect
  */
 @DependsOnGoogleMobileAds
@@ -34,26 +35,28 @@ public fun RewardedAd(
     onLoad: () -> Unit = {}
 ) {
     val ad by rememberRewardedAd(activity)
-    ad.load(
-        adUnitId = adUnitId,
-        onLoad = onLoad,
-        onFailure = onFailure
-    )
-    ad.setListeners(
-        onFailure = onFailure,
-        onDismissed = onDismissed,
-        onShown = onShown,
-        onImpression = onImpression,
-        onClick = onClick
-    )
-    ad.show(
-        onRewardEarned = onRewardEarned
-    )
+    LaunchedEffect(ad){
+        ad.load(
+            adUnitId = adUnitId,
+            onLoad = onLoad,
+            onFailure = onFailure
+        )
+        ad.setListeners(
+            onFailure = onFailure,
+            onDismissed = onDismissed,
+            onShown = onShown,
+            onImpression = onImpression,
+            onClick = onClick
+        )
+        ad.show(
+            onRewardEarned = onRewardEarned
+        )
+    }
 }
 
 /**
  * Displays a pre-loaded Rewarded Ad using a [Composable].
- * @param loadedAd an [RewardedAd] pre-loaded before the call
+ * @param loadedAd an [RewardedAdHandler] pre-loaded before the call
  * @param onRewardEarned Lambda that executes when the user has earned an ad-related reward
  * @param onDismissed Lambda that executes when the user closes the ad
  * @param onShown Lambda expression that executes after the ad is presented
@@ -61,12 +64,12 @@ public fun RewardedAd(
  * @param onClick Lambda expression that executes after the user clicks the ad
  * @param onFailure Lambda expression that executes after the ad fails to load or redirect
  * @see AdUnitId.autoSelect
- * @see RewardedAd.load
+ * @see RewardedAdHandler.load
  */
 @DependsOnGoogleMobileAds
 @Composable
 public fun RewardedAd(
-    loadedAd: RewardedAd,
+    loadedAd: RewardedAdHandler,
     onRewardEarned: () -> Unit,
     onDismissed: () -> Unit = {},
     onShown: () -> Unit = {},
@@ -74,14 +77,16 @@ public fun RewardedAd(
     onClick: () -> Unit = {},
     onFailure: (Exception) -> Unit = {},
 ) {
-    loadedAd.setListeners(
-        onFailure = onFailure,
-        onDismissed = onDismissed,
-        onShown = onShown,
-        onImpression = onImpression,
-        onClick = onClick
-    )
-    loadedAd.show(
-        onRewardEarned = onRewardEarned
-    )
+    LaunchedEffect(loadedAd){
+        loadedAd.setListeners(
+            onFailure = onFailure,
+            onDismissed = onDismissed,
+            onShown = onShown,
+            onImpression = onImpression,
+            onClick = onClick
+        )
+        loadedAd.show(
+            onRewardEarned = onRewardEarned
+        )
+    }
 }
